Show broadcaster song information
When a listener is listening along with someone,
And their currently playing song ends,
It may take up to 5 seconds for them to be resented.

Prior to this commit,
During those (0-5) seconds,
The listen along front end reflects that they’re not listening to anything.

The fact that the listener is not listening to anything _right this second_ is actually the truth, though misleading because they’re still in sync and confusing to see the currently playing song component disappear for 5 seconds and then reappear.

This commit makes a change which causes the front end to display the song the listener _will be synced with_ during those 5 seconds, regardless of whether or not they’ve been synced yet.