Favorite songs while you’re listening along
As a listener,
When I listen along with someone,
And I tap on the song I’m listening to at the bottom of the screen,
Then I see who is listening along with me,
And I see a heart icon.

When I tap the heart icon,
Then the song is added to my library,
And I a popup letting me know the song was added to my library.

—

This commit also widens the scope of our Spotify API privileges to include all API endpoints.

The reason for that is that I don’t want to have to have everyone re-authenticate through the Spotify OAuth web view when I need a new privilege for a new feature.