Turn off repeat on listeners devices
For reasons described in the commit with this SHA (73898a3911707ef6cd69b46f4570dcb3c7e3cf76),
Spotify users who listen along with someone else have their repeat mode set to repeat the current track.

Prior to this commit,
When a listener stops listening along with a broadcaster,
They’re left in an infinite loop playing the last song their broadcaster had been listening to.

This commit turns off repeat on listeners devices when they stop listening along with a broadcaster.

—

This commit also does a necessary refactor of the update_playback_service.

It’s shifted to a more unit tested style because this change made the service extremely difficult to test.

In particular, the test in `unsync_listeners_whose_broadcaster_stopped_broadcasting_spec.rb` was seemingly impossible to write before this refactor.

After the refactor, it was cake.