Log playback service durations
We suspect that as the number of users grow, the time it takes for the first part of this service, `UpdatePlaybackStates.update`, will begin taking a very long time. One network request per user is made in here.

If the duration becomes untenable, we could set up something to do these requests in parallel and wait until they’re all completed before continuing.