Resync out of sync listeners
Given I am listening along with Lucy,
When Lucy scrubs to a new point in the song we’re listening to,
Then my playback is updated to the scrubbed-to point in the song,
So that Lucy and I continue listening to the same music in sync.

Technically, Lucy will have to had scrubbed >= 3 seconds from where she had been listening in the song in order to trigger the resync. A smaller than 3-second difference in playback points is deemed acceptable (and given network latency, anticipated to be around 1-2 seconds in most cases).