Set profile images when users have no email
When a Spotify user has a Spotify profile image, we use that as their avatar.

When a Spotify user has no Spotify profile image, we use the gravatar based on their email address.

Wehn a Spotify user has no email address, we use the gravatar based on their Spotify username.