Do not update users when the rate limit is hit
Setting users to a “not listening” state when the rate limit is reached is causing the user list to flash from a very large list to a very small list frequently.

It also causes listeners to be detached from their broadcasters (basically without fail since the API rate limit is hit so often) leaving them in a repeat-song state.