Fix bug causing listeners to be unhooked
Fix bug causing listeners to be unhooked from their broadcaster.

This happens frequently and immediately.

When a listen along user taps someone else’s name, to listen along with them, they’re synced into the music that the broadcaster is listening to and appear as a listener for the broadcaster, but then they are removed as a listener for that broadcaster almost immediately.

This is because when you listen along with a broadcaster, the listener’s broadcaster is set to the new broadcaster, but their song_uri isn’t updated to the broadcasters song_uri until the request to the spotify API to actually sync up the user with the broadcaster completes, which is a non-negligible (api request) amount of time.

In that time, sometimes the `UnsyncListenersWhoStartedANewSong` command sometimes runs which looks for users with broadcasters who have differing song_uri’s and unhooks them.

This commit makes sure that the song_uri of the new listener is set when the broadcaster is set so that this accidental (sometimes) immediate unhook no longer happens.