Do not open Spotify app when listening to Spotify
We only check to see if users have started listening to Spotify every 10 minutes.

In the mobile app, if it looks like a user is not listening to spotify (they may have started to listen to Spotify in between the every-10-minute checks), and they tap someone else’s name to listen along with them, we open the spotify app to begin Spotify playback.

This commit allows us to check the user’s specific playback status each time they try to listen along with someone else, when initially it looks like they’re not listening to Spotify.