Fix bug causing listener / broadcaster unsyncing
Frequently, when you tap someone else’s name in the listen along app in order to listen along with them, you are synced up to their music, but then the UI indicates that you still listening independently (after flashing, quickly, the correct “you are now a listener” UI).

This is due to a race condition. If someone listens along with another user while the worker process is collecting updated playback states for everyone, the broadcaster attribute written to user’s records is overwritten when the worker finishes it’s job.

This bug occurred pretty frequently when trying to listen along with people. Now it’s fixed!